requirements.txt
altgraph==0.17.2
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==4.1.1
django-import-export==2.9.0
et-xmlfile==1.1.0
filelock==3.8.0
future==0.18.2
MarkupPy==1.14
mysql-connector-python==8.0.29
numpy==1.23.3
odfpy==1.4.1
opencv-python==4.6.0.66
openpyxl==3.0.10
pefile==2022.5.30
Pillow==9.2.0
platformdirs==2.5.2
protobuf==4.21.2
pycryptodome==3.15.0
pycryptodomex==3.15.0
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
pywin32-ctypes==0.2.0
PyYAML==6.0
sqlparse==0.4.3
tablib==3.2.1
tzdata==2022.4
virtualenv==20.16.5
xlrd==2.0.1
xlwt==1.3.0
